separated components and styles into folders
