+++ src/App.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { Switch, Route } from "react-router-dom";
 import { ThemeProvider } from "styled-components";
-import GlobalStyles from "./global";
-import theme from "./theme";
+import GlobalStyles from "./styles/global";
+import theme from "./styles/theme";
 import {
   Navbar,
   Home,
