+++ src/components/Header/Header.jsx
@@ -4,7 +4,7 @@ import { string } from "prop-types";
 const Header = ({ message }) => (
   <>
     <div className="grid-container">
-      <h1>{message}</h1>
+      <h2>{message}</h2>
     </div>
     <div className="grid-container" />
   </>
