+++ src/components/Resume/Resume.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import StyledResume from "./Resume.styled";
-import Header from "./Header";
+import Header from "../Header";
 import Experience from "./Experience";
 import portrait from "../../resources/img/portrait.png";
-import List from "./List";
+import List from "../List";
 import Education from "./Education";
 import ContactInfo from "./ContactInfo";
 import education from "../../resources/data/education.json";
